CompoundLocation half done, created as well as sub_features
This modifies the GenBank/EMBL parser and BioSQL loader to create
the CompoundLocation as part of creating the SeqFeature.

NOTE: Need to examine how the location operator (join, order, etc)
should be stored in BioSQL, and what order is used for the location
parts in BioSQL (by the BioPerl loader at least).